Bluetooth connection and implementation
For this commit, a series of source files have been written to configure and initiate Bluetooth connection to the Arduino using the HC-05.
This commit is a stable implementation to initiate and start a Bluetooth connection which can stream data to a paired Android device. 

For the interface, you can use a Bluetooth terminal application which can be found in Playstore. Remember to pair your mobile device to "FLOURISH" module which is the configure name of the HC-05